Updates for hw_ui script
Move from a regular rotary encoder to the Adafruit seesaw based one.
This one moves the heavy IO (which Pi is bad at) to an M0 controller
which provides the data via I2C.

Also update the display logic to minimize the jitteryness at the
stable play point of buffer = 100%.  It still has a lot of jitter
for small delay values (under 500ms).